backend/requirements.txt
fastapi
uvicorn
openai
llama-cpp-python
requests